feat: add native iOS app via Tauri mobile
Run AgentRove as a real iOS app reusing the existing React frontend.
iOS can't run the local backend sidecar, so mobile is a thin client to
the cloud backend: the desktop sidecar/tray/updater boot paths are gated
to desktop, and the Rust app is restructured into a lib with a mobile
entry point (desktop logic moved to desktop.rs, cfg-gated).

- Cargo: gate desktop-only deps/plugins by target_os (not the custom
  `desktop` cfg, which Cargo's dependency resolver can't evaluate)
- tauri.ios.conf.json: skip sidecar build/bundle, mobile capability,
  development team for signing
- Frontend: build-time platform split (isMobileApp/isDesktopApp) so the
  mobile boot skips the sidecar and points at the cloud via .env.mobile
- iOS safe-area handling so the UI clears the Dynamic Island and home
  indicator (no-op on desktop where the insets are 0)
- Generated Xcode project (gen/apple) with the correct app icon
