+++ frontend/.env.mobile
@@ -1,0 +1,9 @@
+# Mobile (iOS) build env. Loaded by `vite --mode mobile`.
+# VITE_PLATFORM=mobile flips the app to thin-client mode: no local sidecar, talk
+# to a remote backend over the network.
+VITE_PLATFORM=mobile
+
+# Backend the iOS app connects to — mobile is a thin client to the cloud instance.
+# Works from both the simulator and a physical device (https/wss, no localhost).
+VITE_API_BASE_URL=https://claudex.pro/api/v1
+VITE_WS_URL=wss://claudex.pro/api/v1/ws
