+++ frontend/src-tauri/Cargo.toml
@@ -4,21 +4,33 @@ version = "0.1.0"
 description = "Agentrove desktop app"
 edition = "2021"
 
+# Mobile (iOS/Android) links against a static/shared lib and calls `run()`, so the
+# app logic lives in lib.rs; main.rs is a thin desktop entry point.
+[lib]
+name = "agentrove_lib"
+crate-type = ["staticlib", "cdylib", "rlib"]
+
 [build-dependencies]
 tauri-build = { version = "2.5.5", features = [] }
 
 [dependencies]
+tauri = { version = "2.10.2", features = [] }
+tauri-plugin-notification = "2.3.0"
+tauri-plugin-dialog = "2.4.0"
+tauri-plugin-store = "2.4.0"
+tauri-plugin-opener = "2.5.0"
+serde = { version = "1", features = ["derive"] }
+serde_json = "1"
+
+# Desktop-only: the local backend sidecar, tray, updater, and the process plugin
+# don't exist on mobile. Gating by target_os (not the custom `desktop` cfg, which
+# Cargo's dependency resolver can't evaluate) keeps iOS/Android buildable.
+[target.'cfg(not(any(target_os = "ios", target_os = "android")))'.dependencies]
 tauri = { version = "2.10.2", features = ["devtools", "tray-icon"] }
 tauri-plugin-updater = "2.10.0"
 tauri-plugin-process = "2.3.0"
-tauri-plugin-notification = "2.3.0"
-tauri-plugin-dialog = "2.4.0"
-serde = { version = "1", features = ["derive"] }
-serde_json = "1"
 rand = "0.9"
 dirs = "5"
-tauri-plugin-store = "2.4.0"
-tauri-plugin-opener = "2.5.0"
 
 [target.'cfg(target_os = "macos")'.dependencies]
 libc = "0.2"
