+++ frontend/src-tauri/capabilities/mobile.json
@@ -1,0 +1,14 @@
+{
+  "$schema": "../gen/schemas/mobile-schema.json",
+  "identifier": "mobile-capability",
+  "description": "Mobile capability — no window/tray/updater/process plugins on iOS/Android",
+  "platforms": ["iOS", "android"],
+  "windows": ["main"],
+  "permissions": [
+    "core:default",
+    "store:default",
+    "notification:default",
+    "dialog:default",
+    "opener:default"
+  ]
+}
