+++ frontend/src-tauri/gen/apple/.gitignore
@@ -1,0 +1,3 @@
+xcuserdata/
+build/
+Externals/
