+++ frontend/src-tauri/tauri.ios.conf.json
@@ -1,0 +1,19 @@
+{
+  "$schema": "https://schema.tauri.app/config/2",
+  "build": {
+    "beforeDevCommand": "npm run dev:mobile",
+    "beforeBuildCommand": "npm run build:mobile"
+  },
+  "app": {
+    "security": {
+      "capabilities": ["mobile-capability"]
+    }
+  },
+  "bundle": {
+    "resources": [],
+    "createUpdaterArtifacts": false,
+    "iOS": {
+      "developmentTeam": "S4QKFZ7PA6"
+    }
+  }
+}
