+++ frontend/src/styles/globals.css
@@ -166,15 +166,39 @@ body,
   min-height: 100dvh;
 }
 
+/* iOS safe areas: with viewport-fit=cover the webview is edge-to-edge, so inset
+   the app shell by the notch/Dynamic Island (top) and home indicator (bottom).
+   The inset regions show #root's themed background, so they blend in both modes.
+   No-op on desktop/web where the env() values are 0. */
+@supports (padding: env(safe-area-inset-top)) {
+  #root {
+    padding-top: env(safe-area-inset-top);
+    padding-bottom: env(safe-area-inset-bottom);
+    padding-left: env(safe-area-inset-left);
+    padding-right: env(safe-area-inset-right);
+  }
+}
+
 @layer utilities {
+  /* Subtract the safe-area insets that #root now pads with, so full-height
+     containers fit the padded box instead of overflowing it (would scroll on iOS).
+     env() is 0 on desktop/web, so these stay full viewport height there. */
   .h-viewport {
     height: 100vh;
-    height: 100dvh;
+    height: calc(100dvh - env(safe-area-inset-top) - env(safe-area-inset-bottom));
   }
 
   .min-h-viewport {
     min-height: 100vh;
-    min-height: 100dvh;
+    min-height: calc(100dvh - env(safe-area-inset-top) - env(safe-area-inset-bottom));
+  }
+
+  .h-screen {
+    height: calc(100dvh - env(safe-area-inset-top) - env(safe-area-inset-bottom));
+  }
+
+  .min-h-screen {
+    min-height: calc(100dvh - env(safe-area-inset-top) - env(safe-area-inset-bottom));
   }
 
   @keyframes fade-in {
